Fix throw_result for included domains failure
When an included domain resulted in a 'no applicable sender policy'
error, throw_result was being passed a string when  
a symbol was needed ('permerror' vs :permerror). This resulted in a 
'NoMethodError: undefined method `new' for nil:NilClass'
error. Subsequently SPF::Result::PermError was being 
instantiated with 3 args when it was looking for an 
array of args.